2009-03-30  Cory Quammen <cquammen@cs.unc.edu>
   * Types.h: Made 3D float image type actually be floats.

+++ lib/Types.h
@@ -12,7 +12,7 @@ typedef itk::Image<UShortPixelType, Dimension3> UShort3DImageType;
 typedef itk::ImageFileReader<UShort3DImageType> UShort3DImageReaderType;
 typedef itk::VTKImageExport<UShort3DImageType> UShort3DExporterType;
 
-typedef unsigned short FloatPixelType;
+typedef float FloatPixelType;
 typedef itk::Image<FloatPixelType, Dimension3> Float3DImageType;
 typedef itk::ImageFileReader<Float3DImageType> Float3DImageReaderType;
 typedef itk::VTKImageExport<Float3DImageType> Float3DExporterType;
